Fixed runaway score bug and persistent particles bug
Runaway score bug: The score was fine until you missed a target, in which case the score would plummet to zero. Fixed by ensuring that the Miss() function is only called once.

Persistent particles bug: the particle spawners were dependent on the targets to despawn, so if a target was destroyed before the particle spawner it would cause the particle spawner to stay there. Fixed by modifying the existing DestroyWithiTime script and applying it to the particle spawner prefabs, so that they control their own destruction (oddly poetic).